algunos arreglos de alertas

+++ client/src/Components/ServiceCard/ServiceCard.jsx
@@ -3,6 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { getServices, deleteServices } from "../../Redux/Actions";
 import { useParams, useLocation } from "react-router-dom";
 import  Style  from "./ServiceCard.module.css";
+import Swal from "sweetalert2";
 
 const ServiceCard = () => {
   const { id } = useParams();
@@ -12,9 +13,25 @@ const ServiceCard = () => {
   const dispatch = useDispatch();
   const location = useLocation();
 
-  const handleDeleteService = (id) => {
+  const handleDeleteService = async (id) => {
+    const { isConfirmed } = await Swal.fire({
+      title: "¿Estás seguro de que deseas borrar este servicio?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, borrar",
+      cancelButtonText: "Cancelar",
+    });
+    if (isConfirmed) {
     dispatch(deleteServices(id));
-
+    await Swal.fire({
+      title: "Servicio eliminado",
+      icon: "success",
+      text: "Se ha eliminado el servicio.",
+      confirmButtonText: "Aceptar",
+    });
+  };
   };
 
   return (
@@ -27,7 +44,7 @@ const ServiceCard = () => {
             <p>Duración: {service.duration} hs</p>
             <p>Precio: {service.price}</p>
             {location.pathname === `/professionalDetail/${id}` && (
-              <button onClick={() => handleDeleteService(service.id)}>
+              <button className={Style.btnborrar} onClick={() => handleDeleteService(service.id)}>
                 Eliminar
               </button>
             )}
